Monograms
added class nmo:Monogram (as subclass of Thing) : A monogram is a symbol
made by overlapping or combining of letters or other graphemes.
added property nmo:hasMonogram (as subproperty of nmo:hasIconography) :
Identifies the monogram that appears on a numismatic object.

nmo:hasMintmark and nmo:hasControlmark additionally as subpropery of
nmo:hasIconography

added the description for nmo:hasBearsDate