Shape, hasFace ...
New:

nmo:Shape - "The possible shapes numismatic objects might have, could be
standard like round or rectangular or even a polygon."
nmo:hasShape - "Describes the shape of a numismatic object."
nmo:hasFace - "Identifies a particular face of a numismatic object.
Where possible and appropriate, the more specific properties
hasObverse or hasReverse are preferred.”
nmo:hasContext - "Describes the context of a numismatic object."

Updates:
including description for nmo:Find as: "A numismatic object or group of
numismatic objects that was lost or deposited, whether
intentionally or unintentionally, in previous times, and subsequently
recovered. The recovery takes place independently of any
intention of retrieval associated with the original deposition."

Classes DepositionType and DiscoveryType as subclasses of Context.

Deprecate the property: nmo:hasFindType "deprecated - please use
nmo:hasContext instead, togeher with the class Context or its sub
classes: Find, DepositionType and DiscoveryType"