Ontology update
1) Created hasClosingDate and hasOpeningDate
2) hasStartDate and hasEndDate as subproperties of hasDate
3) updated the definition of Hoard
----
Hi Ethan,

perfect, then I will create the two properties (as according sub
properties), here a first proposal for the definitions:
nmo:hasClosingDate: Identifies the latest dating of a numismatic object
within a set of objects. Normally the dating of the youngest coin of a
hoard.

nmo:hasOpeningDate: Identifies the earliest dating of a numismatic
object within a set of objects. Normally the dating of the oldest coin
of a hoard.

I will also change the definition of nmo:Hoard to: A group of two or
more artefacts that have been deliberately brought together, including
at least one numismatic object, and then deposited either accidentally
or deliberately.

Greetings
Karsten